move server-constants.js
