+++ src/lib/notion/getNotionAssetUrls.ts
@@ -1,7 +1,7 @@
 import fetch from "node-fetch";
 import { getError } from "./rpc";
 import { NextApiResponse } from "next";
-import { NOTION_TOKEN, API_ENDPOINT } from "./server-constants";
+import { NOTION_TOKEN, API_ENDPOINT } from "../../constants/notion";
 
 export default async function getNotionAsset(
   res: NextApiResponse,
