+++ src/lib/notion/rpc.ts
@@ -1,5 +1,5 @@
 import fetch, { Response } from "node-fetch";
-import { API_ENDPOINT, NOTION_TOKEN } from "./server-constants";
+import { API_ENDPOINT, NOTION_TOKEN } from "../../constants/notion";
 
 export default async function rpc(fnName: string, body: any) {
   if (!NOTION_TOKEN) {
